Fix svg-text aligning problem
The svg image had a bug. When you add one near (or as in this case, inside) the link tag it is not aligned with the image itself. Applying vertical align property to svg solves this issue. Also the link tag has to be reduced in size or otherwise the link would activate all around the grid row.